refactor(frontend): extract large routes to src/pages/ (Phase 3) (#95)
